Remove float32 logits conversion to save memory
The .float() conversion was creating a 12GB float32 copy of the logits
tensor, causing OOM during RLS training. PyTorch's cross_entropy can
handle bfloat16 inputs directly without numerical issues.

Memory savings with batch_size=96:
- Before: 6.44 GB (bfloat16) + 12.88 GB (float32) = 19.32 GB for logits
- After: 6.44 GB (bfloat16 only)
- Saves: ~12.88 GB

This allows RLS training to fit in 40GB A100 memory.

+++ nanochat/gpt.py
@@ -296,7 +296,8 @@ class GPT(nn.Module):
             # TODO: experiment with Liger Kernels / chunked cross-entropy etc.
             logits = self.lm_head(x)
             logits = softcap * torch.tanh(logits / softcap) # logits softcap
-            logits = logits.float() # use tf32/fp32 for logits
+            # Note: Keeping logits in bfloat16 instead of converting to float32
+            # to save memory. PyTorch cross_entropy handles bfloat16 fine.
             loss = F.cross_entropy(logits.view(-1, logits.size(-1)), targets.view(-1), ignore_index=-1, reduction=loss_reduction)
             if return_state:
                 return loss, final_state
